AArch64: Port Dalvik to 64bit
Change Dalvik such that it compiles and runs as a 64bit program.
Supports only the C interpreter and builds only on 64bit hosts.
The VM runs ordinary Dalvik bytecode. The interpreter has been
changed to have 64bit registers, but the Java types remain
unchanged. The garbage collector, JNI and interpreter are
functional.

+++ vm/Hash.h
@@ -41,7 +41,7 @@ typedef void (*HashFreeFunc)(void* ptr);
 /*
  * Used by dvmHashForeach().
  */
-typedef int (*HashForeachFunc)(void* data, void* arg);
+typedef intptr_t (*HashForeachFunc)(void* data, void* arg);
 
 /*
  * Used by dvmHashForeachRemove().
@@ -156,7 +156,7 @@ bool dvmHashTableRemove(HashTable* pHashTable, u4 hash, void* item);
  *
  * If "func" returns a nonzero value, terminate early and return the value.
  */
-int dvmHashForeach(HashTable* pHashTable, HashForeachFunc func, void* arg);
+intptr_t dvmHashForeach(HashTable* pHashTable, HashForeachFunc func, void* arg);
 
 /*
  * Execute "func" on every entry in the hash table.
